Address PR comments [1]
