+++ packages/relayer/modules/cosmos-to-eth-attested/src/tx_builder.rs
@@ -305,29 +305,3 @@ where
         todo!()
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use attestor_packet_membership::{PacketCommitments, PacketCompact};
-
-    #[test]
-    fn abi_bytes_are_not_json() {
-        // ARRANGE
-        let packets = vec![
-            PacketCompact::new([0x11u8; 32], [0x12u8; 32]),
-            PacketCompact::new([0x21u8; 32], [0x22u8; 32]),
-        ];
-
-        let abi = PacketCommitments::new(packets).to_abi_bytes();
-
-        // ACT
-        let parsed: Result<Vec<Vec<u8>>, _> = serde_json::from_slice(&abi);
-
-        // ASSERT
-        assert!(
-            parsed.is_err(),
-            "ABI-encoded bytes32[] must not be parsed as JSON"
-        );
-    }
-}
