+++ packages/relayer/modules/eth-to-cosmos-attested/Cargo.toml
@@ -29,8 +29,5 @@ attestor-light-client = { workspace = true }
 ibc-proto-eureka          = { workspace = true }
 ibc-eureka-solidity-types = { workspace = true, default-features = true }
 
-[dev-dependencies]
-attestor-packet-membership = { workspace = true }
-
 [build-dependencies]
 tonic-build = { workspace = true }
